# Conflicts:
#	src/modules/workshop/UI/Panel/BottomBar/index.tsx
#	src/modules/workshop/components/Scene/index.tsx

+++ src/modules/workshop/UI/Panel/BottomBar/index.tsx
@@ -7,13 +7,13 @@ import { useModalStore, useProjectStore, useStoreGlobal } from '@/stores/blocks'
 import s from './styles.module.scss'
 
 import useApiInfinite from '@/hooks/useApiInfinite'
-import { getListModularByWallet } from '@/services/api/generative'
+import { getListModularByWallet, handleGetData } from '@/services/api/generative'
 import { TBlockData, TListCurrent } from '@/types'
 import instance from '@/utils/storage/local-storage'
 import { useAppSelector } from '@/stores/hooks'
 import { accountSelector } from '@/stores/states/wallet/selector'
 import { handleConvertData } from '@/utils/convertTraits'
-import { useEffect } from 'react'
+import { useEffect, useMemo, useRef } from 'react'
 
 type TDataFetch = {
   list: TListCurrent
@@ -25,6 +25,7 @@ import { Toaster } from 'react-hot-toast'
 import jsonFile from './mock.json'
 import UnsaveWarningModal from '@/modules/workshop/components/Modal/UnsaveWarningModal'
 import SetProjectNameModal, { SET_PROJECT_NAME_MODAL_ID } from '@/modules/workshop/components/Modal/SetProjectNameModal'
+import { SHA256 } from 'crypto-js'
 
 // const MOCK_ADDRESS = 'bc1p4psqwcglffqz87kl0ynzx26dtxvu3ep75a02d09fshy90awnpewqvkt7er'
 
@@ -56,6 +57,10 @@ export default function BottomBar() {
   const [showUnsaveModal, setShowUnsaveModal] = useState(false)
   const [showSetProjectNameModal, setShowSetProjectNameModal] = useState(false)
 
+  const currentBlockStateRef = useRef(SHA256(JSON.stringify(blocksState)).toString() || '')
+
+  console.log('🚀 ~ BottomBar ~ currentBlockStateRef:', currentBlockStateRef)
+
   const account = useAppSelector(accountSelector)
 
   const id = useId()
@@ -68,6 +73,7 @@ export default function BottomBar() {
     getListModularByWallet,
     {
       ownerAddress: account?.address,
+      // ownerAddress: 'bc1pafhpvjgj5x7era4cv55zdhpl57qvj0c60z084zsl7cwlmn3gq9tq3hqdmn',
       page: 1,
       limit: 20,
     },
@@ -86,17 +92,22 @@ export default function BottomBar() {
     redo()
   }
 
+  const isAllowSave = useMemo(() => {
+    const hashBlockState = SHA256(JSON.stringify(blocksState)).toString()
+
+    return hashBlockState !== currentBlockStateRef.current && blocksState.length > 1
+  }, [blocksState])
+
   const saveAction = async () => {
-    if (blocksState.length < 2 || !blockCurrent || blockCurrent.length === 0) return
+    if (!isAllowSave) return
 
     if (!projectName) {
       openModal({
         id: SET_PROJECT_NAME_MODAL_ID,
-        component: <SetProjectNameModal type="save" />
+        component: <SetProjectNameModal type='save' />,
       })
-      return;
+      return
     }
-
 
     const payload: {
       jsonFile: any
@@ -120,13 +131,13 @@ export default function BottomBar() {
   }
 
   const saveAsAction = async () => {
-    if (blocksState.length < 2 || !blockCurrent || blockCurrent.length === 0) return
+    if (!isAllowSave) return
 
     openModal({
       id: SET_PROJECT_NAME_MODAL_ID,
-      component: <SetProjectNameModal type="save-as" />
+      component: <SetProjectNameModal type='save-as' />,
     })
-    return;
+    return
   }
 
   const loadDataAction = (file) => {
@@ -144,24 +155,23 @@ export default function BottomBar() {
   const handleGetData = async () => {
     const data = (await getListModularByWallet({
       ownerAddress: account?.address,
+      // ownerAddress: 'bc1pafhpvjgj5x7era4cv55zdhpl57qvj0c60z084zsl7cwlmn3gq9tq3hqdmn',
       page: 1,
       limit: 20,
-    })) as TDataFetch
-    const convertData =
-      Array.isArray(data.list) &&
-      data.list.map((item) => {
-        const traits = handleConvertData(item.attributes)
-        return { ...item, traits }
-      })
-    setDataCurrent(convertData)
+    })) as any
+    const listData = data.list as TListCurrent[]
+    return listData
   }
 
-  const handleClickCreateNewProject = () => {
-    // if (blocksState.length > 2) {
-    //   setShowUnsaveModal(true)
-    // }
+  const handleClickCreateNewProject = async () => {
+    if (isAllowSave) {
+      setShowUnsaveModal(true)
+      return
+    }
     createProject()
     deleteAlls()
+    const data = await handleGetData() //reset new project
+    setDataCurrent(data)
   }
 
   useUndoRedoShortcut(undo, redo)
@@ -171,10 +181,6 @@ export default function BottomBar() {
       loadDataAction(renderFile)
     }
   }, [renderFile])
-
-  useEffect(() => {
-    handleGetData()
-  }, [])
 
   // First load, trigger create new project
   useEffect(() => {
@@ -212,10 +218,10 @@ export default function BottomBar() {
         </div>
 
         <div className={s.bottomBar}>
-          <button className={s.bottomBar_btn} onClick={saveAction}>
+          <button className={s.bottomBar_btn} onClick={saveAction} disabled={!isAllowSave}>
             <IcSave /> Save
           </button>
-          <button className={s.bottomBar_btn} onClick={saveAsAction}>
+          <button className={s.bottomBar_btn} onClick={saveAsAction} disabled={!isAllowSave}>
             <IcSave /> Save As
           </button>
           <button className={s.bottomBar_btn} onClick={handleClickCreateNewProject}>
@@ -234,7 +240,7 @@ export default function BottomBar() {
           </button>
         </div>
       </div>
-      {/* <UnsaveWarningModal show={showUnsaveModal} setIsOpen={setShowUnsaveModal} /> */}
+      <UnsaveWarningModal show={showUnsaveModal} setIsOpen={setShowUnsaveModal} />
     </>
   )
 }
